backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.1.18
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.12.1

# Config & auth
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# Data & ML
numpy==1.26.4
joblib==1.3.2
python-dateutil==2.9.0.post0
Pillow


# TensorFlow (CPU)
tensorflow

# Core
pydantic==2.5.2
pydantic-settings==2.1.0
starlette==0.27.0

